+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Splitting inside type parameter and type argument lists (#184).
 * Enforce a blank line before and after classes (#186).
 * Re-indent line doc comments even if they are flush left (#192).
+* More precisely control newlines between declarations (#173).
 
 # 0.1.8